Adds active state styling to nav links
Improves user experience by visually highlighting the active navigation link.

This change introduces active state styling to the navigation links in the dashboard.  It utilizes `next/navigation` to determine the current route and applies a distinct background and text color to the corresponding link.  This provides clear visual feedback to the user, indicating their current location within the application.

Chapter 5
https://nextjs.org/learn/dashboard-app/navigating-between-pages